infernalmobs compat for random enchanted books
im not sure if this should be here or in rlmixins
rltweaker already has a custom override for its own enchantment blacklist

the main issue is that infernalmobs does its own random enchant rolling which works the exact same as vanilla EnchantRandomly loot function, but accesses the whole list of registered enchants.

if the goal is to have SME be late mixin free, this would have to move to rlmixins instead. it does make sense that its here though, as it uses the blacklist defined here.